==> code/0.Extract_Restaurant_Data.ipynb <==
# Applications of NLP in recommender systems

The main goal of the project is to create a recommender system of restaurant based on the reviews of the user. The dataset finally selected is a Yelp Dataset that contains information about different business in the main US and Canada cities as well as complete information about the reviews and users. The dataset in json format can be found here: https://www.yelp.com/dataset 

From all the differents json in the data set, *business* and *reviews* were selected from this project. From the *business* dataset only restaurant were selected and only 10 different cuisines were chosen among the most popular cuisenes.

# Import libraries

import numpy as np
import pandas as pd
from pandas.io.json import json_normalize
import json

## Load Bussiness Data
path = '../data/yelp_academic_dataset_business.json'
business = pd.read_json(path, lines=True)
print(f'The dimensions of the dataset are: {business.shape[0]} rows and {business.shape[1]} columns')

The columns from the dataset *business* contains the following information:

- business_id: *string*, 22 character unique string business id
- name: *string*, the business's name
- address: *string*, the full address of the business
- city: *string*, the city
- state: *string*, the state (if aplicable)
- postal code: *string*, the postal code
- latitude: *float*, latitude
- longitude: *float*, longitude
- stars: *float*, rounded mean of rating
- review_count: *integer*, number of reviews
- is_open: *integer*, 1 is is open
- attributes: *object*, business attributes to values
- hours: an *dictionary* of key day to value hours

df_business = business[business['is_open']==1].drop(['hours','is_open','review_count'], axis=1)
df_explode = df_business.assign(categories = df_business.categories.str.split(', ')).explode('categories')

business.head()

# Select only business == Restaurant
df_business = business[business['is_open']==1].drop(['hours','is_open'], axis=1)
df_explode = df_business.assign(categories = df_business.categories.str.split(', ')).explode('categories')
restaurants = df_explode[df_explode['categories'].str.contains('Restaurants', case=True, na=False)]

# Filter restaurant by cuisine type
cuisine_list = ['Mexican', 'American', 'Japanese', 'Chinese', 'Italian', 'French', 'Indian', 'Mediterranean', 'Thai']
cuisine_restaurants = pd.DataFrame()
for cuisine in cuisine_list:
    data = restaurants[business['categories'].str.contains(cuisine, case=True, na=False)]
    data['categories'] = cuisine
    cuisine_restaurants = pd.concat([cuisine_restaurants, data])

# Remove restaurant that has various types of cuisines
cuisine_restaurants['business_id'].drop_duplicates(inplace=True)
print(f'The dimensions of the dataset are: {cuisine_restaurants.shape[0]} rows' +
      f' and {cuisine_restaurants.shape[1]} columns')

# Export to csv
cuisine_restaurants.to_csv('../data/yelp_restaurants.csv', index=False)

# Load the big Dataset (+6GB) of reviews avoiding the memory error
path = '../data/yelp_academic_dataset_review.json'
size = 50000
review = pd.read_json(path, lines=True,
                      dtype={'review_id':str,'user_id':str,
                             'business_id':str,'stars':int,
                             'date':str,'text':str,'useful':int,
                             'funny':int,'cool':int},
                      chunksize=size)

# Loop over each chunk of data and merge with the bussiness information
chunk_list = []
for chunk_review in review:
    # Drop columns that aren't needed
    chunk_review = chunk_review.drop(['review_id','useful','funny','cool'], axis=1)
    # Renaming column name to avoid conflict with business overall star rating
    chunk_review = chunk_review.rename(columns={'stars': 'review_stars'})
    # Inner merge with edited business file so only reviews related to the business remain
    chunk_merged = pd.merge(chunk_review, cuisine_restaurants, on='business_id', how='inner')
    chunk_list.append(chunk_merged)
    
df = pd.concat(chunk_list, ignore_index=True, join='outer', axis=0)

The columns from the dataset business contains the following information:

- review_id: *string*, 22 character unique review id
- user_id: *string*, 22 character unique user id, maps to the user in user.json
- business_id: *string*, 22 character unique string business id
- stars: *integer*, rating 
- date: *string*, date format
- text: *string*, the review itself
- useful: *integer*, number of useful votes received
- funny: *integer*, number of funny votes received
- cool: *integer*, number of cool votes received

df.head()

# Dimensions of the data
df.shape

# Distribution of rating
df.review_stars.value_counts()

df.to_csv("../data/yelp_reviews_restaurant.csv", index=False)
