==> code/2.Preprocessing.ipynb <==
# 2.- Text Preprocessing using NLP and Calculate Sentiment Analysis 

import pandas as pd
import numpy as np

# Text processing
import spacy
import spacy_lookups_data
import re

# Sentiment Analysis
from textblob import TextBlob

# Scaler
from sklearn.preprocessing import MinMaxScaler

reviews = pd.read_csv('../data/yelp_reviews_restaurant.csv')
reviews.head()

# Remove users that have more than 50 reviews.
grouped_users = reviews.groupby('user_id')['text'].count().reset_index()
grouped_users.head()

result = grouped_users.query('text > 50')
result['text'].sum()

reviews.shape[0]

reviews_filtered = reviews[['user_id', 'name', 'review_stars', 'text']][reviews['user_id'].isin(result['user_id'])]
reviews_filtered.drop_duplicates(inplace=True)
reviews_filtered.shape

## Text preprocessing

# Functions to preprocess text

# convert to lower cases
def to_lower(text):
    return text.lower()

# remove some special characters
def remove_special_chars(text):
    return ''.join(re.sub(r'[-/@"():;^_%&,.!?]', ' ', text))

# remove white spaces
def remove_spaces(text):
    return re.sub(r'\s+',' ', text)

%%time
def clean_text(df):
    df['cleaned_text'] = df['text'].map(to_lower).map(remove_special_chars).map(remove_spaces)
    return df

df = clean_text(reviews_filtered)
df.head()

## Sentiment Analysis

Sentiment analysis is a text analysis method that detects polarity (e.g. a positive or negative opinion) within text, whether a whole document, paragraph, sentence, or clause.

Understanding people’s emotions is essential for businesses since customers are able to express their thoughts and feelings more openly than ever before.

# Calculate sentiment for each review

def sentiment_calc(text):
    try:
        return TextBlob(text).sentiment.polarity
    except:
        return None

df['sentiment'] = df['cleaned_text'].apply(sentiment_calc)

df.reset_index().drop('index', axis=1, inplace=True)

# Check if there are some null values
df.isnull().sum()

## Scale sentiment to be comparable with rating
scaler = MinMaxScaler(feature_range=(1,5))

df['sent_rating'] = scaler.fit_transform(df['sentiment'].values.reshape(-1,1))

df.drop('sentiment', axis=1, inplace=True)

df.head()

# Map user_id with index

df['user_id'].unique(), df['user_id'].nunique()

# Map user_id with index

for i, user_id in enumerate(df['user_id'].unique()):
    idx = df[df['user_id'] == user_id].index
    df.loc[idx,'uid'] = int(i)
    
df['uid'].nunique()

df['uid'] = df['uid'].astype('int16')

df.drop(['user_id', 'text'], axis=1, inplace=True)

df.head()

df.to_csv('../data/reviews_filtered.csv', index=False)
