==> code/3.Models.ipynb <==
# 3. Model selection for recommender systems

import pandas as pd
import numpy as np

#import reader, crossvalidation, models ametrics from surprise
from surprise import Reader, Dataset
from surprise.model_selection import LeaveOneOut, cross_validate
from surprise import KNNBaseline, KNNBasic, SVD, NMF, accuracy

# Load data
df = pd.read_csv('../data/reviews_filtered.csv')
df.head()

# Rating df
df_ratings = df[['uid', 'name', 'review_stars']]
df_ratings.columns = ['uid', 'restaurant', 'rating']
df_ratings.head()

# Sentiment df
df_sentiment = df[['uid', 'name', 'sent_rating']]
df_sentiment.columns = ['uid', 'restaurant', 'rating']
df_sentiment.head()

# Models to test
algorithms_cv = {'svd': SVD(),
              'NMF': NMF(), 
              'kNN_pearson': KNNBasic(sim_options={'name': 'pearson', 'user_based': False }), 
              'kNN_cosine': KNNBasic(sim_options= {'name': 'cosine', 'user_based': False}), 
              'kNN_msd': KNNBasic(sim_options={'name': 'msd', 'user_based': False}), 
              }

def model_validation(df, algorithms):
    
    data = Dataset.load_from_df(df, Reader(rating_scale=(1,5)))

    benchmark = []
    # Iterate over all algorithms
    for model, algorithm in algorithms.items():
        # Perform cross validation
        results = cross_validate(algorithm, data, measures=['RMSE', 'MSE', 'MAE'], cv=10, 
                                 verbose=False)     
        
        # Get results & append algorithm name
        tmp = pd.DataFrame.from_dict(results).mean(axis=0)
        tmp = tmp.append(pd.Series(model, index=['Algorithm']))
        benchmark.append(tmp)
    
    output = pd.DataFrame(benchmark).set_index('Algorithm').sort_values('test_rmse')
    output.columns = ['RMSE(K=10)','MSE(K=10)', 'MAE(K=10)', 'fit_time', 'test_time']
    return output

# Ratings
results_ratings = model_validation(df_ratings, algorithms_cv)
results_ratings

results_ratings.to_csv('../data/model_ratings.csv')

# Sentiment analysis
results_sentiment = model_validation(df_sentiment, algorithms_cv)
results_sentiment

results_sentiment.to_csv('../data/model_sentiment.csv')

# Rounded sentiment
df_sentiment_rounded = df_sentiment.copy()
df_sentiment_rounded['rounded'] = df_sentiment_rounded['rating'].apply(round)
df_sentiment_rounded.drop('rating', axis=1, inplace=True)
df_sentiment_rounded.columns = ['uid', 'restaurant', 'rating']

results_sentiment_rounded = model_validation(df_sentiment_rounded, algorithms_cv)
results_sentiment_rounded

results_sentiment_rounded.to_csv('model_sentiment_rounded.csv')

algorithms_tunning = {'SVD': SVD,
              #'kNN': KNNBasic
              }

#Hyperparameter tunning

from surprise.model_selection import GridSearchCV
def parameter_selection(df, algorithms):
    data = Dataset.load_from_df(df, Reader(rating_scale=(1,5)))
    
    for model, algorithm in algorithms.items():
        if model == 'SVD':
            param_grid = {'n_factors':[50,100,200,500],
                      'n_epochs':[20,30,50], 'lr_all':[0.002,0.005,0.01],
                      'reg_all':[0.02,0.06,0.08]}
        
        elif model == 'kNN':
            param_grid = {'k': [20, 30, 40, 50], #number of neighbors for computation
              'sim_options': {'name': ['msd', 'cosine'],
                              'min_support': [5, 10], #minimum number of common users
                              'user_based': [False]}
              }
        else: raise ValueError('The model provided is not valid')

    gs = GridSearchCV(algorithm, param_grid, measures=['rmse'], cv=2)
    gs.fit(data)
    params = gs.best_params['rmse']
    print(f'The best combination of parameters is {params}')
    return pd.DataFrame.from_dict(gs.cv_results)

parameter_tunning_SVD = parameter_selection(df_ratings, algorithms_tunning)

parameter_tunning_kNN = parameter_selection(df_sentiment, algorithms_tunning)

parameter_tunning_kNN

parameter_tunning_SVD.to_csv('../data/tunning_SVD.csv')
