==> code/4.Review_RS_Evaluation_Metrics.ipynb <==
# Full review based model and evaluation metrics for recommenders

## Preprocessing

# Import libraries
import pandas as pd
import numpy as np
from collections import defaultdict
from functools import reduce

# NLP processing for lemmatizing
import spacy
import spacy_lookups_data

# Fast word2vect embedding
import fse
from fse import IndexedList
import gensim.downloader as api
from fse.models import uSIF

#Recommender
#loaders
from surprise import Dataset, Reader
# cross validation
from surprise.model_selection import cross_validate, KFold, LeaveOneOut
# Models for RS
from surprise import SVD, KNNBasic
# Hypertunning and metrics
from surprise import accuracy
#GridSearchCV

# Other metrics for recommender systems
import recmetrics

# Load dataframe
reviews = pd.read_csv('../data/reviews_filtered.csv')
reviews.drop_duplicates(inplace=True)
print(reviews.shape)
reviews.head()

nlp = spacy.load('en_core_web_sm', disable = ['ner', 'parser']) 
nlp.max_length = 33000000

def get_lemmas(string):
    '''
    This function takes a sentence and returns a clean text
    '''
    doc = nlp(string)
    l_token = [token.lemma_ for token in doc if not token.is_punct | token.is_space | token.is_digit | 
               token.like_url | token.like_num | token.like_email & token.is_oov]
    return ' '.join(l_token)

# Lemmatize the words of each sentence
reviews['cleaned_text'] = reviews['cleaned_text'].apply(get_lemmas)
reviews.head()

# Tokenize (split by words)
sentences = reviews['cleaned_text'].apply(lambda x: x.split(' ')).to_list()

# Load word dictionary and apply it to create a model
glove = api.load("glove-wiki-gigaword-100")

model = uSIF(glove, workers=2, lang_freq="en")

# Train the fse model
s = IndexedList(sentences)

model.train(s)

matrix = []
for i in range(reviews.shape[0]):
    matrix.append(model.sv[i])
vector_df = pd.DataFrame(matrix)
del matrix

vector_df.to_csv('../data/vector_df.csv')

## Recommendations from models

### Custom functions in order to process and calculate performance metrics

# Function to get the index of all comments

def get_indices(df, threshold=3.5):
    '''
    Retrieve the index of the most rated item for each user(row)
    '''
    uids = []
    idx = []
    ratings = []
    df = df.sort_values('rating', ascending=False)
    for i in df.uid.unique():
        idx.append(df[(df.uid == i) & (df.rating > threshold)].index.to_list())
        ratings.append(df['rating'][(df.uid == i) & (df.rating > threshold)].to_list())
        uids.append(i)
        
    return pd.DataFrame({'user_id': uids, 'idx_actual': idx, 'ratings': ratings})

#Get recommendations based on rating:

def get_users_predictions(uid, df, r_column='rating', k=10):
    '''
    This functions takes an user(uid) from a dataframe(df) and returns a
    list the most k (int) rated items.
    '''
    recommended_items = df[['uid', r_column]][df['uid'] == uid]
    recommended_items = recommended_items.sort_values(r_column, ascending=False).head(k)  
    return recommended_items


def similar_comments(lst, k=50):
    '''
    This functions takes a list of indices and returns the most similar comments, 
    based on the model previously calculated before. The model computes the distances between vector 
    of words of each comments using cosine similarity.
    '''
    idx_comments = []
    for i in lst:
        try:
            similar = [idx for idx in model.sv.most_similar(i)]
            for idx in similar:
                if idx not in idx_comments and idx[0] not in lst:
                    idx_comments.append(idx)
        except KeyError: # The index may has been eliminated
            continue
    # Filter by the similarity
    idx_comments.sort(key=lambda x:x[1], reverse=True)
    # Get only the k indices of comments
    result = [i[0] for i in idx_comments][:k]
    return result

def get_restaurant_name(lst, df=reviews):
    '''
    Given a list of idx return the restaurant name
    '''
    result = []
    for i in lst:
        try:
            result.append(reviews.loc[i,'name'])
        except:
            result.append(np.nan)
    return result

def get_ratings(lst, df=reviews):
    '''
    Given a list of idx return the rating
    '''
    result = []
    for i in lst:
        try:
            result.append(reviews.loc[i,'review_stars'])
        except:
            result.append(np.nan)
    return result

df_ratings = reviews[['uid', 'name', 'review_stars']]
df_ratings.columns = ['uid', 'rid', 'rating'] #uid = user_id / rid = restaurant

# Get indices of the commments most rated for each user
rec_comments = get_indices(df_ratings)
rec_comments.head()

# Drop this rating column because the rating that we want is from the list of similar comments 
rec_comments.drop('ratings', axis=1, inplace=True)

# Get indices of the predictions
rec_comments['idx_rec'] = rec_comments['idx_actual'].apply(similar_comments)
rec_comments.head()

# Get the name of the restaurants
rec_comments['pred_com'] = rec_comments['idx_rec'].apply(get_ratings)
#rec_comments['pred_com'] = rec_comments['idx_rec'].apply(get_restaurant_name)
rec_comments.head()

ark(rec_comments.loc[10,'pred_com'], rec_comments.loc[10,'idx_rec'])

### List of recommendations for rating only

data_rating = Dataset.load_from_df(df_ratings, Reader(rating_scale=(1,5))) 

# Optimized Model
svd = SVD(n_factors = 50, n_epochs = 50, lr_all = 0.002, reg_all =  0.08)
trainset = data_rating.build_full_trainset()
svd.fit(trainset)

# Map iid to name restaurant
name_iid = {}
for i in trainset.ir:
    name_iid[trainset.to_raw_iid(i)] = i

predictions = {}
for i in range(df_ratings.shape[0]):
    try:
        predictions[i] = svd.predict(df_ratings.loc[i, 'uid'], df_ratings.loc[i,'rid']).est
    except:
        predictions[i] = np.nan

df_ratings['pred'] = predictions.values()

df_rating_stars = df_ratings[['uid', 'rid', 'pred']]
df_rating_stars.columns = ['uid', 'rid', 'rating']
df_rating_stars

# Index of the recommendations as well as the rated give by the user
rec_ratings = get_indices(df_rating_stars)
rec_ratings.head()

rec_ratings['rat_actual'] = rec_ratings['idx_actual'].apply(get_ratings)
rec_ratings

### Sentiment analysis based recommendation

df_sent = reviews[['uid', 'name', 'sent_rating']]
df_sent.columns = ['uid', 'rid', 'rating'] #uid = user_id / rid = restaurant
data_sent = Dataset.load_from_df(df_sent, Reader(rating_scale=(1,5)))

# Fine tunned model for sentiment analysis
kNN = KNNBasic(k=50, min_k=5, sim_options = {'name': 'msd', 'user_based': [False]})
trainset = data_rating.build_full_trainset()
kNN.fit(trainset)

predictions = {}
for i in range(df_sent.shape[0]):
    try:
        predictions[i] = kNN.predict(df_sent.loc[i, 'uid'], df_sent.loc[i,'rid']).est
    except:
        predictions[i] = np.nan

df_sent['pred'] = predictions.values()

df_sent = df_sent[['uid', 'rid', 'pred']]
df_sent.columns = ['uid', 'rid', 'rating']
df_sent

rec_sentiment = get_indices(df_sent)
#rec_sentiment['sent_pred'] = rec_ratings['idx_actual'].apply(get_restaurant_name)
rec_sentiment.head()

rec_sentiment['rat_sent_actual'] = rec_sentiment['idx_actual'].apply(get_ratings)
rec_sentiment

###  Create a matrix with the list of all predictions for all users

# Drop unnecessary columns
rec_comments.drop('idx_actual', axis=1, inplace=True)
rec_ratings.drop('idx_actual', axis=1, inplace=True)
rec_sentiment.drop('idx_actual', axis=1, inplace=True)

dfs = [rec_comments, rec_ratings, rec_sentiment]
recommendations = reduce(lambda left,right: pd.merge(left,right,on='user_id'), dfs)

recommendations.drop(['ratings_x', 'ratings_y'], axis=1, inplace=True)

recommendations.head()

### Comparation of performance between recommenders

#### Average Precision at K

def apk(rating, pred, k=10):
    '''
    This functions calculates the precision at Kth metric for each user in the dataset.
    '''
    score = 0
    n_hits = 0
    
    for r,p in (zip(rating, pred)):
        if len(pred) > k:
            pred = pred[:k]
            rating = rating[:k]
            
        for i, p in enumerate(pred):
            # If the recommendation is relevant add it to the result
            if rating[i] >= 3.5:
                n_hits +=1
                score += n_hits / (i+1)
            
    
        if not rating:
            return np.nan

        return score / min(len(pred), k)
    

def ark(rating, pred, k=10):
    '''
    This functions calculates the recall at Kth metric for each user in the dataset.
    '''
    
    n_hits = 0
    
    for a,p in (zip(rating, pred)):
        if len(pred) > k:
            pred = pred[:k]
            rating = rating[:k]
    
        for i, p in enumerate(pred):
            if rating[i] > 3.5:
                n_hits +=1
    
        if not rating:
            return np.nan
        
        return n_hits / min(len(rating), k)
    
    

# Calculate the average precission and recall for each user

for i in range(recommendations.shape[0]):
    
    # Average Precision at K
    recommendations.loc[i, 'apk_comments'] = apk(recommendations.loc[i,'pred_com'], recommendations.loc[i,'idx_rec'])
    recommendations.loc[i, 'apk_rating'] = apk(recommendations.loc[i,'rat_actual'], recommendations.loc[i,'idx_rec'])
    recommendations.loc[i, 'apk_sentiment'] = apk(recommendations.loc[i,'rat_sent_actual'], recommendations.loc[i,'idx_rec'])
    
    # Average Recall at K
    recommendations.loc[i, 'ark_comments'] = ark(recommendations.loc[i,'pred_com'], recommendations.loc[i,'idx_rec'])
    recommendations.loc[i, 'ark_rating'] = ark(recommendations.loc[i,'rat_actual'], recommendations.loc[i,'idx_rec'])
    recommendations.loc[i, 'ark_sentiment'] = ark(recommendations.loc[i,'rat_sent_actual'], recommendations.loc[i,'idx_rec'])


## Calculate the global metric of mean average precision and recall for each dataset

mapk_comments = np.mean(recommendations.apk_comments)
mapk_rating = np.mean(recommendations.apk_rating)
mapk_sentiment = np.mean(recommendations.apk_sentiment)

mark_comments = np.mean(recommendations.ark_comments)
mark_rating = np.mean(recommendations.ark_rating)
mark_sentiment = np.mean(recommendations.ark_sentiment)

print(mapk_comments, mapk_rating, mapk_sentiment)
print(mark_comments, mark_rating, mark_sentiment)

print('#### Mean Average Precission ####')
print(f'The Mean Average Precission at 10 (MAP@10) for rating dataset is: {round(mapk_rating,3)}')
print(f'The Mean Average Precission at 10 (MAP@10) for sentiment dataset is: {round(mapk_sentiment,3)}')
print(f'The Mean Average Precission at 10 (MAP@10) for comments dataset is: {round(mapk_comments,3)}')
print('\n')
print('#### Mean Average Recall ####')
print(f'The Mean Average Recall at 10 (MAR@10) for rating dataset is: {round(mark_rating,3)}')
print(f'The Mean Average Recall at 10 (MAR@10) for sentiment dataset is: {round(mark_sentiment,3)}')
print(f'The Mean Average Recall at 10 (MAR@10) for comments dataset is: {round(mark_comments,3)}')
